src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "cachetools>=7.0.6",
    "cohere>=5.21.1",
    "fastapi>=0.135.3",
    "firecrawl-py>=4.23.0",
    "langgraph>=1.1.9",
    "openai>=2.32.0",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlglot>=30.6.0",
    "uvicorn>=0.44.0",
]
